% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cross_counts.R
\name{add_cross_counts}
\alias{add_cross_counts}
\title{add_cross_counts}
\usage{
add_cross_counts(df, group_by_vars)
}
\arguments{
\item{df}{TODO}

\item{group_by_vars}{TODO}
}
\value{
TODO
}
\description{
add_cross_counts
}
