% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset_layout.R
\name{create_dataset_layout}
\alias{create_dataset_layout}
\title{create_dataset_layout}
\usage{
create_dataset_layout(full_df)
}
\arguments{
\item{full_df}{TODO}
}
\value{
TODO
}
\description{
create_dataset_layout
}
