% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset_layout.R
\name{create_dataset_layout_keep}
\alias{create_dataset_layout_keep}
\title{create_dataset_layout_keep}
\usage{
create_dataset_layout_keep(full_df)
}
\arguments{
\item{full_df}{TODO}
}
\value{
TODO
}
\description{
create_dataset_layout_keep
}
