% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_initial_list.R
\name{create_file_structure}
\alias{create_file_structure}
\title{create_initial_list}
\usage{
create_file_structure(jaw_e)
}
\arguments{
\item{jaw_e}{TODO}
}
\value{
TODO
}
\description{
create_initial_list
}
