% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_initial_list.R
\name{create_initial_list}
\alias{create_initial_list}
\title{create_initial_list}
\usage{
create_initial_list(args)
}
\arguments{
\item{args}{TODO}
}
\value{
TODO
}
\description{
create_initial_list
}
