% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_session_list.R
\name{create_session_list}
\alias{create_session_list}
\title{create_session_list}
\usage{
create_session_list(totem, sas_file_path)
}
\arguments{
\item{totem}{TODO}

\item{sas_file_path}{TODO}
}
\value{
TODO
}
\description{
create_session_list
}
