% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_meta.R
\name{df_meta_contents}
\alias{df_meta_contents}
\title{df_meta_contents}
\usage{
df_meta_contents(df, contents)
}
\arguments{
\item{df}{TODO}

\item{contents}{TODO}
}
\value{
TODO
}
\description{
df_meta_contents
}
