% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__add_before_filter.R
\name{e__add_before_filter}
\alias{e__add_before_filter}
\title{e__add_before_filter}
\usage{
e__add_before_filter(session_name, current_row, exclude = F, outer_env = totem)
}
\arguments{
\item{session_name}{TODO}

\item{current_row}{TODO}

\item{exclude}{TODO}

\item{outer_env}{TODO}
}
\value{
TODO
}
\description{
e__add_before_filter
}
