% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__add_before_filter.R
\name{e__add_before_filter_full_data_column}
\alias{e__add_before_filter_full_data_column}
\title{e__add_before_filter_full_data_column}
\usage{
e__add_before_filter_full_data_column(
  session_name,
  current_row,
  df_obj,
  exclude = F,
  outer_env = totem
)
}
\arguments{
\item{session_name}{TODO}

\item{current_row}{TODO}

\item{df_obj}{TODO}

\item{exclude}{TODO}

\item{outer_env}{TODO}
}
\value{
TODO
}
\description{
e__add_before_filter_full_data_column
}
