% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__add_column_label.R
\name{e__add_column_label}
\alias{e__add_column_label}
\title{e__add_column_label}
\usage{
e__add_column_label(
  treeviewcolumn,
  label,
  j,
  var_class = NULL,
  tool_tip = NULL,
  sec_label,
  outer_env = totem,
  obj_env = inner_env
)
}
\arguments{
\item{treeviewcolumn}{TODO}

\item{label}{TODO}

\item{j}{TODO}

\item{var_class}{TODO}

\item{tool_tip}{TODO}

\item{sec_label}{TODO}

\item{outer_env}{TODO}

\item{obj_env}{TODO}
}
\value{
TODO
}
\description{
e__add_column_label
}
