% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__add_before_filter.R
\name{e__add_count_to_df_summary}
\alias{e__add_count_to_df_summary}
\title{e__add_count_to_df_summary}
\usage{
e__add_count_to_df_summary(session_name, cross_tab_names, outer_env = totem)
}
\arguments{
\item{session_name}{TODO}

\item{cross_tab_names}{TODO}

\item{outer_env}{TODO}
}
\value{
TODO
}
\description{
e__add_count_to_df_summary
}
