% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__add_session.R
\name{e__add_session}
\alias{e__add_session}
\title{e__add_session}
\usage{
e__add_session(sas_file_path, outer_env = totem)
}
\arguments{
\item{sas_file_path}{TODO}

\item{outer_env}{TODO}
}
\value{
TODO
}
\description{
e__add_session
}
