% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__all_event_functions.R
\name{e__all_event_functions}
\alias{e__all_event_functions}
\title{e__all_event_functions}
\usage{
e__all_event_functions(outer_env = totem)
}
\arguments{
\item{outer_env}{TODO}
}
\value{
TODO
}
\description{
e__all_event_functions
}
