% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__all_menu_events.R
\name{e__all_menu_events}
\alias{e__all_menu_events}
\title{e__all_menu_events}
\usage{
e__all_menu_events(mi, data, ...)
}
\arguments{
\item{mi}{TODO}

\item{data}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
e__all_menu_events
}
