% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__close_all_windows.R
\name{e__close_all_windows}
\alias{e__close_all_windows}
\title{e__close_all_windows}
\usage{
e__close_all_windows(session_name, outer_env = totem)
}
\arguments{
\item{session_name}{TODO}

\item{outer_env}{TODO}
}
\value{
TODO
}
\description{
e__close_all_windows
}
