% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__code_r_add.R
\name{e__code_r_add}
\alias{e__code_r_add}
\title{e__code_r_add}
\usage{
e__code_r_add(session_name, outer_env = totem)
}
\arguments{
\item{session_name}{TODO}

\item{outer_env}{TODO}
}
\value{
TODO
}
\description{
e__code_r_add
}
