% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__code_r_add.R
\name{e__code_r_add_cmd}
\alias{e__code_r_add_cmd}
\title{e__code_r_add_cmd}
\usage{
e__code_r_add_cmd(session_name, cmd, outer_env = totem)
}
\arguments{
\item{session_name}{TODO}

\item{cmd}{TODO}

\item{outer_env}{TODO}
}
\value{
TODO
}
\description{
e__code_r_add_cmd
}
