% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__copy_if_then.R
\name{e__copy_if_then_do}
\alias{e__copy_if_then_do}
\title{e__copy_if_then_do}
\usage{
e__copy_if_then_do(session_name, current_row, df_obj, outer_env = totem)
}
\arguments{
\item{session_name}{TODO}

\item{current_row}{TODO}

\item{df_obj}{TODO}

\item{outer_env}{TODO}
}
\value{
TODO
}
\description{
e__copy_if_then_do
}
