% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__create_settings.R
\name{e__create_settings}
\alias{e__create_settings}
\title{e__create_settings}
\usage{
e__create_settings(outer_env = totem)
}
\arguments{
\item{outer_env}{TODO}
}
\value{
TODO
}
\description{
e__create_settings
}
