% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__df_tree.R
\name{e__df_tree}
\alias{e__df_tree}
\title{e__df_tree}
\usage{
e__df_tree(
  session_name,
  passed_box,
  rows_length,
  event_mapping = NULL,
  style_list = NULL,
  is_value_table = F,
  is_meta_table = F,
  is_data_code_table = F,
  is_file_history_table = F,
  is_full_data_table = F,
  outer_env = totem
)
}
\arguments{
\item{session_name}{TODO}

\item{passed_box}{TODO}

\item{rows_length}{TODO}

\item{event_mapping}{TODO}

\item{style_list}{TODO}

\item{is_value_table}{TODO}

\item{is_meta_table}{TODO}

\item{is_data_code_table}{TODO}

\item{is_file_history_table}{TODO}

\item{is_full_data_table}{TODO}

\item{outer_env}{TODO}
}
\value{
TODO
}
\description{
e__df_tree
}
