% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__df_view.R
\name{e__df_view}
\alias{e__df_view}
\title{e__df_view}
\usage{
e__df_view(df, title, height = 300, width = 500, outer_env = totem)
}
\arguments{
\item{df}{TODO}

\item{title}{TODO}

\item{height}{TODO}

\item{width}{TODO}

\item{outer_env}{TODO}
}
\value{
TODO
}
\description{
e__df_view
}
