% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__generic_filter_function.R
\name{e__generic_filter_function}
\alias{e__generic_filter_function}
\title{e__generic_filter_function}
\usage{
e__generic_filter_function(
  e,
  df,
  u__order_by,
  u__filter,
  u__select,
  table_obj,
  outer_env = totem
)
}
\arguments{
\item{e}{TODO}

\item{df}{TODO}

\item{u__order_by}{TODO}

\item{u__filter}{TODO}

\item{u__select}{TODO}

\item{table_obj}{TODO}

\item{outer_env}{TODO}
}
\value{
TODO
}
\description{
e__generic_filter_function
}
