% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__inverted_view.R
\name{e__inverted_view}
\alias{e__inverted_view}
\title{e__inverted_view}
\usage{
e__inverted_view(session_name, current_row, outer_env = totem)
}
\arguments{
\item{session_name}{TODO}

\item{current_row}{TODO}

\item{outer_env}{TODO}
}
\value{
TODO
}
\description{
e__inverted_view
}
