% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__load_dataset.R
\name{e__load_dataset}
\alias{e__load_dataset}
\title{e__load_dataset}
\usage{
e__load_dataset(session_name, outer_env = totem)
}
\arguments{
\item{session_name}{TODO}

\item{outer_env}{TODO}
}
\value{
TODO
}
\description{
e__load_dataset
}
