% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__load_dataset.R
\name{e__load_dataset_filter_inner_select}
\alias{e__load_dataset_filter_inner_select}
\title{e__load_dataset_filter_inner_select}
\usage{
e__load_dataset_filter_inner_select(session_name, df, outer_env = totem)
}
\arguments{
\item{session_name}{TODO}

\item{df}{TODO}

\item{outer_env}{TODO}
}
\value{
TODO
}
\description{
e__load_dataset_filter_inner_select
}
