% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__new_tree_view_column.R
\name{e__new_tree_view_column}
\alias{e__new_tree_view_column}
\title{e__new_tree_view_column}
\usage{
e__new_tree_view_column(df, j, outer_env = totem, obj_env = inner_env)
}
\arguments{
\item{df}{TODO}

\item{j}{TODO}

\item{outer_env}{TODO}

\item{obj_env}{TODO}
}
\value{
TODO
}
\description{
e__new_tree_view_column
}
