% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__order_by_setup.R
\name{e__order_by_setup}
\alias{e__order_by_setup}
\title{e__order_by_setup}
\usage{
e__order_by_setup(header_table, row_1, obj_env = inner_env)
}
\arguments{
\item{header_table}{TODO}

\item{row_1}{TODO}

\item{obj_env}{TODO}
}
\value{
TODO
}
\description{
e__order_by_setup
}
