% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__add_before_filter.R
\name{e__set_before_code}
\alias{e__set_before_code}
\title{e__set_before_code}
\usage{
e__set_before_code(session_name, cmd, outer_env = totem)
}
\arguments{
\item{session_name}{TODO}

\item{cmd}{TODO}

\item{outer_env}{TODO}
}
\value{
TODO
}
\description{
e__set_before_code
}
