% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_start.R
\name{e__start}
\alias{e__start}
\title{e__start}
\usage{
e__start(sas_file_path, outer_env = totem, assign_env = .GlobalEnv)
}
\arguments{
\item{sas_file_path}{TODO}

\item{outer_env}{TODO}

\item{assign_env}{TODO}
}
\value{
TODO
}
\description{
e__start
}
