% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__table_cell_events.R
\name{e__table_cell_events}
\alias{e__table_cell_events}
\title{e__table_cell_events}
\usage{
e__table_cell_events(
  event,
  row.idx,
  col.idx,
  outer_env = totem,
  obj_env = inner_env
)
}
\arguments{
\item{event}{TODO}

\item{row.idx}{TODO}

\item{col.idx}{TODO}

\item{outer_env}{TODO}

\item{obj_env}{TODO}
}
\value{
TODO
}
\description{
e__table_cell_events
}
