% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__table_obj_function.R
\name{e__table_obj_function}
\alias{e__table_obj_function}
\title{e__table_obj_function}
\usage{
e__table_obj_function(box, outer_env = totem, obj_env = inner_env)
}
\arguments{
\item{box}{TODO}

\item{outer_env}{TODO}

\item{obj_env}{TODO}
}
\value{
TODO
}
\description{
e__table_obj_function
}
