% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__tree_view_column.R
\name{e__tree_view_column_btn_press}
\alias{e__tree_view_column_btn_press}
\title{e__tree_view_column_btn_press}
\usage{
e__tree_view_column_btn_press(widget, event, data)
}
\arguments{
\item{widget}{TODO}

\item{event}{TODO}

\item{data}{TODO}
}
\value{
TODO
}
\description{
e__tree_view_column_btn_press
}
