% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas_contents.R
\name{sas_contents}
\alias{sas_contents}
\title{Sas Contents v0.0.1}
\usage{
sas_contents(sas_path)
}
\arguments{
\item{sas_path}{Path to a .sas7bdat file.}
}
\value{
TODO data.frame with the following columns:
variable - names of the variables in the dataset.
length - the storage length of the variable.
type - what kind of variable.
label - the variable label.
n - the row count.

Note the function will always return a data.frame. If it fails, then the row count will be zero.
}
\description{
Sas Contents v0.0.1
}
