% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u__button.R
\name{u__button}
\alias{u__button}
\title{u__button}
\usage{
u__button(
  box,
  start = T,
  padding = 5,
  stock_id = NULL,
  but_txt = NULL,
  tool_tip = NULL,
  call_back_fct = NULL,
  data = NULL
)
}
\arguments{
\item{box}{TODO}

\item{start}{TODO}

\item{padding}{TODO}

\item{stock_id}{TODO}

\item{but_txt}{TODO}

\item{tool_tip}{TODO}

\item{call_back_fct}{TODO}

\item{data}{TODO}
}
\value{
TODO
}
\description{
u__button
}
