% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e__tree_view_column.R
\name{z__tree_view_column_get_col_idx}
\alias{z__tree_view_column_get_col_idx}
\title{z__tree_view_column_get_col_idx}
\usage{
z__tree_view_column_get_col_idx(column)
}
\arguments{
\item{column}{TODO}
}
\value{
TODO
}
\description{
z__tree_view_column_get_col_idx
}
